requirements.txt
# Core Django
Django==5.1.3
djangorestframework==3.15.2
gunicorn==23.0.0
whitenoise==6.8.2

# Database
psycopg==3.2.3
dj-database-url

# Task Queue & Caching
redis==5.2.0
celery==5.4.0

# Utilities
setuptools==74.1.2
requests

# Blog System
django-markdownx==4.0.7
markdown==3.7
Pillow==11.0.0
python-slugify==8.0.4
beautifulsoup4==4.12.3
